--- a/Common/Exceptions/InvalidTokenPythonExceptionInterpreter.cs
+++ b/Common/Exceptions/InvalidTokenPythonExceptionInterpreter.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
  *
@@ -36,7 +36,6 @@ namespace QuantConnect.Exceptions
         public override bool CanInterpret(Exception exception)
         {
             return base.CanInterpret(exception) &&
-                exception.Message.Contains("SyntaxError") &&
                 exception.Message.Contains("invalid token");
         }