… Tests/Common/Data/CalendarConsolidatorsTests.cs
… Tests/Common/Data/Custom/PythonCustomDataTests.cs
… Tests/Common/Data/SliceTests.cs
… Tests/Common/Exceptions/InvalidTokenPythonExceptionInterpreterTests.cs
… Tests/Common/Notifications/NotificationManagerTests.cs
… Tests/Common/Orders/Fees/BackwardsCompatibilityFeeModelTests.cs
… Tests/Common/Orders/Fills/BackwardsCompatibilityFillModelsTests.cs
… Tests/Common/Securities/DynamicSecurityDataTests.cs
… Tests/Common/Util/ExtensionsTests.cs
… Tests/Indicators/PythonIndicatorNoinheritanceTests.cs
… Tests/Indicators/PythonIndicatorNoinheritanceTestsLegacy.cs
… Tests/Indicators/PythonIndicatorTests.cs
… Tests/Python/AlgorithmPythonWrapperTests.cs
… Tests/Python/DataConsolidatorPythonWrapperTests.cs
… Tests/Python/MethodOverloadTests.cs
… Tests/Python/NamedArgumentsTests.cs
… Tests/Python/PandasConverterTests.BackwardsCompatibility.cs
… Tests/Python/PandasConverterTests.cs
… Tests/Python/PortfolioCustomModelTests.cs
… Tests/Python/PythonOptionTests.cs
… Tests/Python/PythonPackagesTests.cs
… Tests/Python/PythonWrapperTests.cs
… Tests/Python/SecurityCustomModelTests.cs
… Tests/QuantConnect.Tests.csproj
… Tests/Research/QuantBookFundamentalTests.cs
… Tests/ResearchRegressionTests.cs