<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Bump pythonNet version 2.0.12
- Updates after pythonNet rebase
#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
See https://github.com/QuantConnect/pythonnet/pull/64
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Update
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Running tests, research.


#### Memory sanity test
Running `CoarseFineFundamentalComboAlgorithm` using `self.__numberOfSymbols = 50;self.__numberOfSymbolsFine = 10`
`Master`
![image](https://user-images.githubusercontent.com/18473240/165858936-e73b44bd-476f-4fa5-acde-d8ec941b30d8.png)

`PR`
![image](https://user-images.githubusercontent.com/18473240/165859093-cd2026b6-13a6-45ef-82df-62d1fe0ba729.png)


#### Performance sanity test:
```
HistoryRequestBenchmark PR
49.38 seconds at 37k data points per second. Processing total of 1,822,036 data points.
48.48 seconds at 38k data points per second. Processing total of 1,822,036 data points.
47.68 seconds at 38k data points per second. Processing total of 1,822,036 data points.
```
```
BasicTemplateBenchmark PR
51.53 seconds at 51k data points per second. Processing total of 2,639,319 data points.
51.72 seconds at 51k data points per second. Processing total of 2,639,319 data points.
50.72 seconds at 52k data points per second. Processing total of 2,639,319 data points.
```
```
BasicTemplateBenchmark MASTER
48.77 seconds at 54k data points per second. Processing total of 2,639,319 data points.
46.56 seconds at 57k data points per second. Processing total of 2,639,319 data points.
47.14 seconds at 56k data points per second. Processing total of 2,639,319 data points.
```
```
HistoryRequestBenchmark MASTER
46.75 seconds at 39k data points per second. Processing total of 1,822,036 data points.
47.13 seconds at 39k data points per second. Processing total of 1,822,036 data points.
48.57 seconds at 38k data points per second. Processing total of 1,822,036 data points.
```
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] Refactor (non-breaking change which improves implementation)
- [ ] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->